Fix error handling during option parsing and exit properly after --help and --version
After Electron is fully initialized, it seems that process.exit() is executed as a delayed command (via the event queue), probalby to enable BrowserWindows to trigger 'unload' and similar events. This leeds to code after process.exit() being executed even in case of failure. Simply return from the current method to the event queue often does the trick, but that's not so easy in case of nested functions like option parsing in yargs. But we can utilize exceptions for these cases.