Fix fullscreen support for systems with plain X11 but no window manager
Chromium has an issue with fullscreen on plain X11. The window doesn’t cover the whole screen, but leaves a 1px wide line at the bottom and right. This fix sets minimum window and content size explicitly to fit the whole screen if the kiosk browser is started in fullscreen mode. Unfortunately, this method breaks the possiblity to leave fullscreen afterwards on some systems.