Process command line options and settings before app is fully initialized
Electron-settings is supposed to be only used after the app has triggered the 'ready' event and therefore we could only process the command line arguments and settings after that event. The side effect was that some Chrome command line switches had no effect because they were set to late in the initialization process. This commit ensures that the user data directory of the app is actually present and then just uses electron-settings before the 'ready' event. Even though this procedure is discouraged by the authors of electron-settings, it seems to just work fine. The existence of the user data directory seam to be the only real prerequisite when using electron-settings.