Remove --testapp and default of -l/--url in favor of special kiosk://testapp and kiosk://home urls
The behaviour of the various CLI switches was quite entangled and not intuitive. Now the internal pages index.html and testapp.html can be accessed via a special kiosk:// protocol schema, similar to chrome:// that is used in Chromium. Defaults for -l/--url have been removed because it was entangled with the positional url argument. This also made it difficult to properly figure out which url to load when the builtin web server is enabled. Now the behavior for the url is:
- kiosk://home is used if NO URL is supplied and the builtin SERVER is OFF
- file://absolute_path_of_url is used if URL is SUPPLIED and the builtin SERVER is OFF
- http://localhost:port/index.html is used if NO URL is supplied and the builtin SERVER is ON
- http://localhost:port/url is used if URL is SUPPLIED and the buildin SERVER is ON
This should fit most needs and feels quite intuitive to me.