Add --cover-displays option that allows to maximize the browser window across several displays on supported platforms
The displays are provides via a comma separated list of display numbers (which, unfortunately, might be platform-dependent). This option can also be used to send a window to a certain display before setting it to fullscreen. Certain platforms like e.g. macOS Mojave do not support windows that span multipe displays.