Warn when workArea is smaller than the display bounds
When the browser window is configured to cover several displays, but the work area of a display is smaller then its bounds, the result of setting the window size and position is very specific to the used window/desktop manager on Linux. Therefore, it is best to use this options with window/desktop managers that don't contain any additional UI elements that reduce the work area of a window (like openbox).