Deny more than one instance per app name
The Chromium process model is build around a single main process. Using
several main processes ("instances") that access the same user data
leads to race conditions and crashes which are hard to debug and can't
be reliably prevented by the user.

Therefore, each instances must now set a different app name through the
`--app-name-suffix` CLI option. If a second instance with the same name
is detected, the app will exit.